requirements.txt
clearml==1.6.4
omegaconf==2.2.2
hydra-core==1.2.0
matplotlib==3.5.2
networkx==2.8.1
tensorflow==2.7.0
tqdm==4.64.0
pyyaml==6.0
joblib==1.1.0
pandas==1.4.3
numpy==1.21.5
scikit-learn==1.1.1
